Wrote RSpec tests for tutees_controller, tutors_controller, and application_controller
Wrote tests to ensure that the update and validate params functions in tutees_controller
and tutors_controller are functional. Wrote tests for application_controller to ensure that
the after_sign_in_path_for function works. Wrote test for the tutor_match function for tutees
to see if there is no available tutor, none will be assigned. Wrote test for tutee_match function
for tutors_controller to see if there is no assigned tutees, the tutees array will be empty.